requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
distro==1.4.0
Django==3.1
django-allauth==0.43.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-formtools==2.2
django-rest-auth==0.9.5
djangorestframework==3.11.1
filelock==3.0.12
Geocoding==1.4.3
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
KdQuery==0.2.2
lockfile==0.12.2
msgpack==0.6.2
numpy==1.19.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
pipenv==2020.8.13
progress==1.5
psycopg2-binary==2.8.6
pycountry==20.7.3
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
webencodings==0.5.1
